pyproject.toml
[project]
name = "cat-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.134.0",
    "google-api-core>=2.25.2",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "line-bot-sdk>=3.22.0",
    "notion-client>=3.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
